+++ agrest/src/main/java/io/agrest/ResourceEntity.java
@@ -39,6 +39,7 @@ public class ResourceEntity<T> {
     private AgRelationship incoming;
     private List<Ordering> orderings;
     private Expression qualifier;
+    private Map<String, EntityProperty> includedExtraProperties;
     private Map<String, EntityProperty> extraProperties;
     private int fetchOffset;
     private int fetchLimit;
@@ -51,6 +52,7 @@ public class ResourceEntity<T> {
         this.children = new HashMap<>();
         this.orderings = new ArrayList<>(2);
         this.extraProperties = new HashMap<>();
+        this.includedExtraProperties = new HashMap<>();
         this.agEntity = agEntity;
     }
 
@@ -130,6 +132,10 @@ public class ResourceEntity<T> {
 
     public Map<String, EntityProperty> getExtraProperties() {
         return extraProperties;
+    }
+
+    public Map<String, EntityProperty> getIncludedExtraProperties() {
+        return includedExtraProperties;
     }
 
     public boolean isIdIncluded() {