+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 Release 3.1
 
 * #358 Protocol Enhancement: 'include\exclude' parameters should take an array of values 
+* #392 Request include/exclude settings are ignored for dynamic request properties
 
 Release 3.0